Incorporate entity_confusion_matrix as a first-class citizen
