+++ tests/shared/hardware_utils.py
@@ -3,6 +3,7 @@ import csv
 import logging
 from builtins import range
 
+from fonduer.learning.utils import entity_confusion_matrix
 from fonduer.supervision.models import GoldLabel, GoldLabelKey
 
 try:
@@ -114,17 +115,6 @@ def load_hardware_labels(
     session.commit()
 
     logger.info(f"GoldLabels created: {labels}")
-
-
-def entity_confusion_matrix(pred, gold):
-    if not isinstance(pred, set):
-        pred = set(pred)
-    if not isinstance(gold, set):
-        gold = set(gold)
-    TP = pred.intersection(gold)
-    FP = pred.difference(gold)
-    FN = gold.difference(pred)
-    return (TP, FP, FN)
 
 
 def entity_level_f1(
